requirements.txt
streamlit==1.35.0
tensorflow==2.18.0
numpy==2.0.2
pillow==11.2.1
matplotlib==3.10.0
